+++ tuxemon/event/actions/char_wander.py
@@ -12,7 +12,7 @@ from tuxemon.event import get_npc
 from tuxemon.event.eventaction import EventAction
 from tuxemon.map import get_coords, get_direction
 from tuxemon.session import Session
-from tuxemon.states.world.worldstate import WorldState
+from tuxemon.states.world_state import WorldState
 
 logger = logging.getLogger(__name__)
