+++ tuxemon/event/actions/play_map_animation.py
@@ -10,7 +10,7 @@ from tuxemon.animation_entity import setup_and_play_animation
 from tuxemon.event import get_npc
 from tuxemon.event.eventaction import EventAction
 from tuxemon.session import Session
-from tuxemon.states.world.worldstate import WorldState
+from tuxemon.states.world_state import WorldState
 
 logger = logging.getLogger(__name__)
