+++ tuxemon/event/actions/screen_transition.py
@@ -9,7 +9,7 @@ from tuxemon.event.eventaction import EventAction
 from tuxemon.graphics import ColorLike, string_to_colorlike
 from tuxemon.prepare import BLACK_COLOR, TRANS_TIME
 from tuxemon.session import Session
-from tuxemon.states.world.worldstate import WorldState
+from tuxemon.states.world_state import WorldState
 
 
 @final
