+++ tuxemon/event/actions/start_cinema_mode.py
@@ -7,7 +7,7 @@ from typing import Optional, final
 
 from tuxemon.event.eventaction import EventAction
 from tuxemon.session import Session
-from tuxemon.states.world.worldstate import WorldState
+from tuxemon.states.world_state import WorldState
 
 
 @final
