--- tuxemon/states/combat/__init__.py
@@ -1,3 +1,0 @@
-# SPDX-License-Identifier: GPL-3.0
-# Copyright (c) 2014-2025 William Edwards <shadowapex@gmail.com>, Benjamin Bean <superman2k5@gmail.com>
-from .combat_menus import CombatTargetMenuState, MainCombatMenuState
